# src/main/java/com/nexym/clinic/domain/availability/model/Availability.java
package com.nexym.clinic.domain.availability.model;

import lombok.Builder;
import lombok.Data;

import java.io.Serializable;
import java.time.LocalDateTime;

@Data
@Builder
public class Availability implements Serializable {

    private Long id;
    private LocalDateTime startDate;
    private LocalDateTime endDate;
    private LocalDateTime creationDate;
    private LocalDateTime modifiedDate;
}